requirements.txt
# Core dependencies
asyncio
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncpg>=0.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration and data handling
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Text processing and analysis
nltk>=3.8.0
textstat>=0.7.0
langdetect>=1.0.9

# Monitoring and metrics
psutil>=5.9.0
prometheus-client>=0.17.0

# Utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional visualization
matplotlib>=3.7.0
plotly>=5.15.0
pandas>=2.0.0